+++ packages/components/Epub/converters/wax.js
@@ -1,3 +1,5 @@
+const hljs = require('highlight.js')
+
 module.exports = (
   $,
   fragmentTitle,
@@ -44,6 +46,8 @@ module.exports = (
       .attr('class', 'chapter-number')
       .html(fragmentNumber)
       .appendTo(innerContainer)
+  } else {
+    // if unumbered component
   }
   $('<h1/>')
     .attr('class', 'ct')
@@ -58,6 +62,17 @@ module.exports = (
     )
 
     $elem.replaceWith(blockquote)
+  }
+
+  const replaceWithPre = className => (i, elem) => {
+    const $elem = $(elem)
+    const { source } = $elem[0].attribs
+    const { language } = $elem[0].attribs
+
+    const highLighter = hljs.highlight(language, source)
+    const pre = $(`<pre class="${language}"/>`).append(highLighter.value)
+
+    $elem.replaceWith(pre)
   }
 
   const replaceWithText = (i, elem) => {
@@ -86,6 +101,18 @@ module.exports = (
     $elem.replaceWith(list)
   }
 
+  const replaceWithFigure = className => (i, elem) => {
+    const $elem = $(elem)
+
+    const figure = $(
+      `<figure><img src="${$elem[0].attribs.src}"/><figcaption>${
+        $elem[0].attribs.figcaption
+      }</figcaption></figure>`,
+    )
+
+    $elem.replaceWith(figure)
+  }
+
   // add namespaces
   $('html').attr({
     xmlns: 'http://www.w3.org/1999/xhtml',
@@ -106,6 +133,9 @@ module.exports = (
   $('source-note').each(replaceWithParagraph('exsn'))
   $('ol[styling="qa"]').each(replaceWithList('di'))
   $('ol[styling="unstyled"]').each(replaceWithList('none'))
+
+  $('figure').each(replaceWithFigure(''))
+  $('script').each(replaceWithPre('pre'))
 
   // remove "uploads" from the start of each src attribute
   $('[src]').each((i, elem) => {