+++ packages/components/xpub-submit/src/components/Notes.js
@@ -2,9 +2,10 @@ import React from 'react'
 import { branch, renderNothing } from 'recompose'
 import { FormSection } from 'redux-form'
 import { NoteEditor } from 'xpub-edit'
+import { withJournal } from 'xpub-journal'
 import { ValidatedField } from '@pubsweet/ui'
 import { required } from 'xpub-validators'
-import { Section } from '../styles'
+import { Section, SubNote } from '../styles'
 
 // TODO: this is only here because prosemirror would save the title in the
 // metadata as html instead of plain text. we need to maybe find a better
@@ -15,42 +16,36 @@ const stripHtml = htmlString => {
   return temp.textContent
 }
 
-const FundingInput = input => (
-  <NoteEditor
-    placeholder="Enter an acknowledgment…"
-    title="Funding body acknowledgement (required)"
-    {...input}
-  />
-)
+const Input = extraProps => input => <NoteEditor {...input} {...extraProps} />
 
-const InstructionsInput = input => (
-  <NoteEditor
-    placeholder="Enter instructions for the editor…"
-    title="Special instructions (confidential, to Editors only)"
-    {...input}
-  />
-)
-
-const Notes = ({ readonly }) => (
+const Notes = ({ readonly, journal }) => (
   <FormSection name="notes">
     <Section id="notes.fundingAcknowledgement">
       <ValidatedField
-        component={FundingInput}
+        component={Input(journal.notes.fundingAcknowledgement)}
         name="fundingAcknowledgement"
         readonly={readonly}
         validate={[required]}
       />
+      <SubNote>
+        {stripHtml(journal.notes.fundingAcknowledgement.description)}
+      </SubNote>
     </Section>
 
     <Section id="notes.specialInstructions">
       <ValidatedField
-        component={InstructionsInput}
+        component={Input(journal.notes.specialInstructions)}
         name="specialInstructions"
         parse={stripHtml}
         readonly={readonly}
       />
+      <SubNote>
+        {stripHtml(journal.notes.specialInstructions.description)}
+      </SubNote>
     </Section>
   </FormSection>
 )
 
-export default branch(({ readonly }) => readonly === true, renderNothing)(Notes)
+export default withJournal(
+  branch(({ readonly }) => readonly === true, renderNothing)(Notes),
+)