+++ packages/components/xpub-submit/src/components/SupplementaryFiles.js
@@ -1,24 +1,26 @@
 import React from 'react'
 import { branch, renderComponent } from 'recompose'
 import { FormSection } from 'redux-form'
+import { withJournal } from 'xpub-journal'
+import ReactHtmlParser from 'react-html-parser'
 import { Supplementary, ValidatedField, Attachment } from '@pubsweet/ui'
-import { Section, Legend } from '../styles'
+import { Section, Legend, SubNote } from '../styles'
 
 const FileInput = uploadFile => ({ value, ...input }) => (
   <Supplementary files={value} uploadFile={uploadFile} {...input} />
 )
 
-const SupplementaryFiles = ({ uploadFile, readonly }) => (
+const SupplementaryFiles = ({ uploadFile, readonly, journal }) => (
   <FormSection name="files">
     <Section id="files.supplementary">
       <Legend htmlFor="supplementary">Upload supplementary materials</Legend>
-
       <ValidatedField
         component={FileInput(uploadFile)}
         name="supplementary"
         readonly={readonly}
       />
     </Section>
+    <SubNote>{ReactHtmlParser(journal.supplementary.description)} </SubNote>
   </FormSection>
 )
 
@@ -35,7 +37,9 @@ const SupplementaryFilesNonEditable = ({ readonly, version }) => (
   </Section>
 )
 
-export default branch(
-  ({ readonly }) => readonly === true,
-  renderComponent(SupplementaryFilesNonEditable),
-)(SupplementaryFiles)
+export default withJournal(
+  branch(
+    ({ readonly }) => readonly === true,
+    renderComponent(SupplementaryFilesNonEditable),
+  )(SupplementaryFiles),
+)