+++ packages/styleguide/styleguide.config.js
@@ -54,4 +54,5 @@ module.exports = {
     transforms: { dangerousTaggedTemplateString: true },
   },
   pagePerSection: true,
+  require: ['babel-polyfill'],
 }