+++ packages/components/xpub-edit/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.1"></a>
+## [1.0.1](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/xpub-edit@1.0.0...xpub-edit@1.0.1) (2018-06-19)
+
+
+
+
+**Note:** Version bump only for package xpub-edit
+
 <a name="1.0.0"></a>
 # [1.0.0](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/xpub-edit@0.1.5...xpub-edit@1.0.0) (2018-06-01)